pyproject.toml
[tool.poetry]
name = "sc-server-auth"
version = "0.1.0"
description = ""
authors = ["Shtress <stas.butrin@mail.ru>"]
license = "MIT License"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "0.75.0"
pydantic = "1.7.4"
PyJWT = "1.7.1"
pyOpenSSL = "22.0.0"
sqlalchemy = "1.4.31"
uvicorn = "0.15.0"
requests = "2.27.1"
websocket-client = "1.3.2"
black = "^22.6.0"
pylint = "2.13.7"
isort = "^5.10.1"
coverage = "^6.4.1"
psycopg2-binary = "^2.9.3"
toml = "^0.10.2"
python-dotenv = "^0.20.0"
google-api-python-client = "^2.53.0"
google-auth-oauthlib = "^0.5.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
